requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes
-e .
accelerate==1.10.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
argilla==2.8.0
async-timeout==4.0.3
attrs==25.3.0
authlib==1.6.1
backoff==2.2.1
bcrypt==4.3.0
bitsandbytes==0.47.0
blinker==1.9.0
blip-ci==0.0.5
build==1.3.0
cachecontrol==0.14.3
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.3
chromadb==1.0.17
click==8.2.1
clickhouse-connect==0.5.25
clip-interrogator==0.5.4
colorama==0.4.6 ; (os_name != 'nt' and sys_platform == 'win32') or (os_name == 'nt' and sys_platform != 'darwin' and sys_platform != 'linux')
coloredlogs==15.0.1
cryptography==45.0.6
cython==3.1.3
dataclasses-json==0.6.7
datasets==4.0.0
deprecated==1.2.18
dill==0.3.7
distro==1.9.0
docstring-parser==0.17.0
dparse==0.6.4
durationpy==0.10
et-xmlfile==2.0.0
exceptiongroup==1.3.0
fairscale==0.4.4
filelock==3.16.1
filetype==1.2.0
firebase-admin==6.8.0
flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.3.0
ftfy==6.3.1
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.73.0
google-cloud-bigquery==3.35.1
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0 ; platform_python_implementation != 'PyPy'
google-cloud-resource-manager==1.14.2
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-texttospeech==2.27.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.4 ; platform_machine == 'AMD64' or platform_machine == 'WIN32' or platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'ppc64le' or platform_machine == 'win32' or platform_machine == 'x86_64'
grpc-google-iam-v1==0.14.2
grpcio==1.74.0
grpcio-status==1.62.3
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.7 ; platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'arm64' or platform_machine == 'x86_64'
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
importlib-metadata==8.7.0
importlib-resources==6.5.2
itsdangerous==2.2.0
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kubernetes==33.1.0
langchain==0.3.27
langchain-chroma==0.2.5
langchain-community==0.3.27
langchain-core==0.3.74
langchain-google-genai==2.1.9
langchain-google-vertexai==2.0.9
langchain-huggingface==0.3.1
langchain-text-splitters==0.3.9
langdetect==1.0.9
langsmith==0.4.14
lightning-utilities==0.15.2
lxml==6.0.0
lz4==4.4.4
markdown==3.8.2
markdown-it-py==4.0.0
markupsafe==3.0.2
marshmallow==3.26.1
marshmallow-enum==1.5.1
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
msg-parser==1.2.0
msgpack==1.1.1
multidict==6.6.4
multiprocess==0.70.15
mypy-extensions==1.1.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1 ; platform_machine == 'x86_64' and sys_platform == 'linux'
nvidia-cuda-cupti-cu12==12.8.90 ; platform_machine == 'x86_64' and sys_platform == 'linux'
nvidia-cuda-nvrtc-cu12==12.8.93 ; platform_machine == 'x86_64' and sys_platform == 'linux'
nvidia-cuda-runtime-cu12==12.8.90 ; platform_machine == 'x86_64' and sys_platform == 'linux'
nvidia-cudnn-cu12==9.10.2.21 ; platform_machine == 'x86_64' and sys_platform == 'linux'
nvidia-cufft-cu12==11.3.3.83 ; platform_machine == 'x86_64' and sys_platform == 'linux'
nvidia-cufile-cu12==1.13.1.3 ; platform_machine == 'x86_64' and sys_platform == 'linux'
nvidia-curand-cu12==10.3.9.90 ; platform_machine == 'x86_64' and sys_platform == 'linux'
nvidia-cusolver-cu12==11.7.3.90 ; platform_machine == 'x86_64' and sys_platform == 'linux'
nvidia-cusparse-cu12==12.5.8.93 ; platform_machine == 'x86_64' and sys_platform == 'linux'
nvidia-cusparselt-cu12==0.7.1 ; platform_machine == 'x86_64' and sys_platform == 'linux'
nvidia-nccl-cu12==2.27.3 ; platform_machine == 'x86_64' and sys_platform == 'linux'
nvidia-nvjitlink-cu12==12.8.93 ; platform_machine == 'x86_64' and sys_platform == 'linux'
nvidia-nvtx-cu12==12.8.90 ; platform_machine == 'x86_64' and sys_platform == 'linux'
oauthlib==3.3.1
olefile==0.47
onnxruntime==1.22.1
open-clip-torch==2.32.0
openai==1.99.9
openpyxl==3.1.5
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.15.0
opentelemetry-proto==1.15.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.2
overrides==7.7.0
packaging==24.2
pandas==2.3.1
pdf2image==1.17.0
pdfminer-six==20221105
pillow==11.3.0
posthog==5.4.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==4.25.8
psutil==6.1.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pybase64==1.4.2
pycparser==2.22
pydantic==2.9.2
pydantic-core==2.23.4
pydantic-settings==2.10.1
pygments==2.19.2
pyjwt==2.10.1
pypandoc==1.15
pyparsing==3.2.3
pypika==0.48.9
pyproject-hooks==1.2.0
pyreadline3==3.5.4 ; sys_platform == 'win32'
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-magic==0.4.27
python-pptx==1.0.2
pytorch-lightning==2.5.3
pytz==2025.2
pyyaml==6.0.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.0
rsa==4.9.1
ruamel-yaml==0.18.14
ruamel-yaml-clib==0.2.12 ; platform_python_implementation == 'CPython'
safetensors==0.6.2
safety==3.6.0
safety-schemas==0.0.14
scikit-learn==1.7.1
scipy==1.15.3
sentence-transformers==5.1.0
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.43
standardwebhooks==1.0.0
sympy==1.14.0
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.6.0
timm==0.4.12
tokenizers==0.21.4
tomli==2.2.1
tomlkit==0.13.3
torch==2.8.0
torchmetrics==1.8.1
torchvision==0.23.0
tqdm==4.67.1
transformers==4.55.2
triton==3.4.0 ; platform_machine == 'x86_64' and sys_platform == 'linux'
typer==0.16.0
types-deprecated==1.2.15.20250304
types-python-dateutil==2.9.0.20250809
typing-extensions==4.14.1
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
unstructured==0.6.11
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0 ; platform_python_implementation != 'PyPy' and sys_platform != 'cygwin' and sys_platform != 'win32'
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
werkzeug==3.1.3
wrapt==1.17.3
xlrd==2.0.2
xlsxwriter==3.2.5
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
